% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof-tree.R
\name{prof.tree}
\alias{prof.tree}
\title{Summarise Output of R Sampling Profiler}
\usage{
prof.tree(filename = "Rprof.out")
}
\arguments{
\item{filename}{Name of a file produced by Rprof().}
}
\description{
Summarise the output of the Rprof function to show the amount of time used by different R functions.
}
\examples{
\dontrun{
Rprof()
some code to be profiled
Rprof(NULL)
prof.tree()
}

}

