% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof-tree.R
\name{prof.tree}
\alias{prof.tree}
\alias{print.ProfTree}
\title{Summarise Output of R Sampling Profiler}
\usage{
prof.tree(filename = "Rprof.out")

\method{print}{ProfTree}(x, limit = 25, ...)
}
\arguments{
\item{filename}{Name of a file produced by \code{Rprof()}.}

\item{x}{A \code{ProfTree} object.}

\item{limit}{The maximum number of nodes to print. Can be \code{NULL} if the entire tree should be printed.}

\item{...}{not used.}
}
\value{
An object of \code{ProfTree} and \code{Node} classes.
}
\description{
Summarise the output of the \code{\link{Rprof}} function to show the amount of time used by different R functions as tree structure.
}
\examples{
Rprof(tmp <- tempfile())
example(glm)
Rprof(NULL)
tree <- prof.tree(tmp)
print(tree, limit = 20)
unlink(tmp)

}
\seealso{
\code{\link{Rprof}} \code{\link{summaryRprof}} \code{\link[data.tree]{plot.Node}}
}
